Largest prime number

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -6,8 +6,8 @@
  */
 int main(void)
 {
-unsigned long int prime = 612852475143;
-	unsigned long int i = 3;
+	int prime = 612852475143;
+	int i = 3;
 
 	while (i <= prime / 2)
 		if (prime % i == 0)
